Fix in the ping-pong example protocol transition.
Problem: when transitioning states, the server sends Version than GameStarted, in sequence. Depending on the network distance and settings,
both messages arrive at the same time and the client sees the 2 messages, while still in the PreGame state, and the latter is denied, as
it can only be receive after the transition from the PreGame to the Game state.

Solution: do not send messages in succession. The "Ping-Pong" aspect is specially important in the transitions. While we may tweak the base implementation, it would require a lot of optimizations to go away. Therefore, we are currently tweaking the protocol and enforcing "the right way" of transitioning.

Implementation: when implementing, we found out that the server was missing the code to verify the client protocol version so, in the end, by simply avoiding the client to omit an answer in the pre-game state was enough to fix the issues.